refact:[step4] view 클래스 분리 리팩토링
