+++ src/main/java/step3_racing_car/RacingGame.java
@@ -8,8 +8,8 @@ public class RacingGame {
     }
     void playGame() {
         // 게임 세팅
-        List<String> participantCarNames = InputOutputView.scanCarNames();
-        int roundCount = InputOutputView.scanRoundCount();
+        List<String> participantCarNames = InputView.scanCarNames();
+        int roundCount = InputView.scanRoundCount();
 
         // 게임 룰 세팅 후 라운드 진행 - Random 전략으로 이동한다.
         RacingCars racingCars = new RacingCars(participantCarNames);
@@ -17,7 +17,7 @@ public class RacingGame {
         gameRound.progressAllRound(roundCount, new RandomStrategy());
 
         // 우승자 선정
-        InputOutputView.printGameWinner(prizeWinners(racingCars));
+        OutputView.printGameWinner(prizeWinners(racingCars));
     }
 
     private RacingCars prizeWinners(RacingCars racingCars){
